requirements.txt
feedparser==6.0.11
newspaper3k==0.2.8
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
beautifulsoup4==4.12.2
lxml==4.9.3
pillow==10.2.0
google-generativeai==0.8.3
sqlalchemy==2.0.25
tenacity==8.2.3
aiohttp==3.9.1
flask==3.0.0
flask-cors==4.0.0
psutil==5.9.6